+++ packages/cmd/__tests__/CommandProcessor.test.ts
@@ -689,7 +689,7 @@ describe("Command Processor", () => {
 
         let error;
         try {
-            const validateResponse: ICommandResponse = await processor.invoke({arguments: undefined});
+            const validateResponse: ICommandResponse = await processor.invoke({ arguments: undefined });
         } catch (e) {
             error = e;
         }
@@ -712,7 +712,7 @@ describe("Command Processor", () => {
 
         let error;
         try {
-            const parms: any = {arguments: {_: [], $0: ""}, responseFormat: "blah", silent: true};
+            const parms: any = { arguments: { _: [], $0: "" }, responseFormat: "blah", silent: true };
             const validateResponse: ICommandResponse = await processor.invoke(parms);
         } catch (e) {
             error = e;
@@ -736,7 +736,7 @@ describe("Command Processor", () => {
 
         let error;
         try {
-            const parms: any = {arguments: {_: undefined, $0: ""}, responseFormat: "json", silent: true};
+            const parms: any = { arguments: { _: undefined, $0: "" }, responseFormat: "json", silent: true };
             const validateResponse: ICommandResponse = await processor.invoke(parms);
         } catch (e) {
             error = e;
@@ -758,7 +758,7 @@ describe("Command Processor", () => {
             promptPhrase: "dummydummy"
         });
 
-        const parms: any = {arguments: {_: ["banana"], $0: "", valid: false}, responseFormat: "json", silent: true};
+        const parms: any = { arguments: { _: ["banana"], $0: "", valid: false }, responseFormat: "json", silent: true };
         const commandResponse: ICommandResponse = await processor.invoke(parms);
 
         expect(commandResponse).toBeDefined();
@@ -783,7 +783,7 @@ describe("Command Processor", () => {
         });
 
         const parms: any = {
-            arguments: {_: ["check", "for", "banana"], $0: "", valid: false},
+            arguments: { _: ["check", "for", "banana"], $0: "", valid: false },
             responseFormat: "json", silent: true
         };
         const commandResponse: ICommandResponse = await processor.invoke(parms);
@@ -810,7 +810,7 @@ describe("Command Processor", () => {
         });
 
         const parms: any = {
-            arguments: {_: ["check", "for", "banana"], $0: "", syntaxThrow: true},
+            arguments: { _: ["check", "for", "banana"], $0: "", syntaxThrow: true },
             responseFormat: "json", silent: true
         };
         const commandResponse: ICommandResponse = await processor.invoke(parms);
@@ -832,7 +832,7 @@ describe("Command Processor", () => {
             promptPhrase: "dummydummy"
         });
 
-        const parms: any = {arguments: {_: [], $0: "", syntaxThrow: true}, responseFormat: "json", silent: true};
+        const parms: any = { arguments: { _: [], $0: "", syntaxThrow: true }, responseFormat: "json", silent: true };
         const commandResponse: ICommandResponse = await processor.invoke(parms);
 
         expect(commandResponse).toBeDefined();
@@ -867,7 +867,7 @@ describe("Command Processor", () => {
             })
         });
 
-        const parms: any = {arguments: {_: [], $0: "", syntaxThrow: true}, responseFormat: "json", silent: true};
+        const parms: any = { arguments: { _: [], $0: "", syntaxThrow: true }, responseFormat: "json", silent: true };
         const commandResponse: ICommandResponse = await processor.invoke(parms);
 
         expect(mockLogInfo).toHaveBeenCalled();
@@ -894,11 +894,11 @@ describe("Command Processor", () => {
 
         // Mock the profile loader
         CommandProfileLoader.loader = jest.fn((args) => {
-            throw new ImperativeError({msg: "Profile loading failed!"});
+            throw new ImperativeError({ msg: "Profile loading failed!" });
         });
 
         const parms: any = {
-            arguments: {_: ["check", "for", "banana"], $0: "", valid: true},
+            arguments: { _: ["check", "for", "banana"], $0: "", valid: true },
             responseFormat: "json", silent: true
         };
         const commandResponse: ICommandResponse = await processor.invoke(parms);
@@ -1372,6 +1372,311 @@ describe("Command Processor", () => {
         expect(commandResponse).toMatchSnapshot();
     });
 
+    it("should display input value for simple parm when --show-inputs-only flag is set", async () => {
+
+        // values to test
+        const parm1Key = `parm1`;
+        const parm1Value = `value1`;
+
+        // Allocate the command processor
+        const processor: CommandProcessor = new CommandProcessor({
+            envVariablePrefix: ENV_VAR_PREFIX,
+            fullDefinition: SAMPLE_COMPLEX_COMMAND, // `group action`
+            definition: { // `object`
+                name: "banana",
+                description: "The banana command",
+                type: "command",
+                handler: __dirname + "/__model__/TestCmdHandler",
+                options: [
+                    {
+                        name: parm1Key,
+                        type: "string",
+                        description: "The first parameter",
+                    }
+                ],
+            },
+            helpGenerator: FAKE_HELP_GENERATOR,
+            profileManagerFactory: FAKE_PROFILE_MANAGER_FACTORY,
+            rootCommandName: SAMPLE_ROOT_COMMAND,
+            commandLine: "",
+            promptPhrase: "dummydummy"
+        });
+
+        const parms: any = {
+            arguments: {
+                _: ["check", "for", "banana"],
+                $0: "",
+                [parm1Key]: parm1Value,
+                valid: true,
+                showInputsOnly: true,
+            },
+            silent: true
+        };
+        const commandResponse: ICommandResponse = await processor.invoke(parms);
+        expect(commandResponse.data.commandValues[parm1Key]).toBe(parm1Value);
+    });
+
+    it("should display input value for simple parm when --show-inputs-only flag is set and v1 profile exists", async () => {
+
+        // values to test
+        const parm1Key = `parm1`;
+        const parm1Value = `value1`;
+
+        // Allocate the command processor
+        const processor: CommandProcessor = new CommandProcessor({
+            envVariablePrefix: ENV_VAR_PREFIX,
+            fullDefinition: SAMPLE_CMD_WITH_OPTS_AND_PROF, // `group action`
+            definition: { // `object`
+                name: "banana",
+                description: "The banana command",
+                type: "command",
+                handler: __dirname + "/__model__/TestCmdHandler",
+                options: [
+                    {
+                        name: parm1Key,
+                        type: "string",
+                        description: "The first parameter",
+                    }
+                ],
+            },
+            helpGenerator: FAKE_HELP_GENERATOR,
+            profileManagerFactory: FAKE_PROFILE_MANAGER_FACTORY,
+            rootCommandName: SAMPLE_ROOT_COMMAND,
+            commandLine: "",
+            promptPhrase: "dummydummy"
+        });
+
+        // Mock the profile loader
+        (CommandProfileLoader.loader as any) = jest.fn((args) => {
+            return {
+                loadProfiles: (profArgs: any) => {
+                    return;
+                }
+            };
+        });
+
+        // return the "fake" args object with values from profile
+        CliUtils.getOptValueFromProfiles = jest.fn((cmdProfiles, profileDef, allOpts) => {
+            return {
+                color: "yellow"
+            };
+        });
+
+        const parms: any = {
+            arguments: {
+                _: ["check", "for", "banana"],
+                $0: "",
+                [parm1Key]: parm1Value,
+                valid: true,
+                showInputsOnly: true,
+            },
+            silent: true
+        };
+        const commandResponse: ICommandResponse = await processor.invoke(parms);
+        expect(commandResponse.data.locations[0]).toContain('home');
+        expect(commandResponse.data.profileVersion).toBe('v1');
+    });
+
+    it("should display input value for simple parm when --show-inputs-only flag is set and v2 config exists", async () => {
+
+        // values to test
+        const parm1Key = `parm1`;
+        const parm1Value = `value1`;
+
+        await setupConfigToLoad({
+            "profiles": {
+                "fruit": {
+                    "properties": {
+                        "origin": "California"
+                    },
+                    "profiles": {
+                        "apple": {
+                            "type": "fruit",
+                            "properties": {
+                                "color": "red"
+                            }
+                        },
+                        "banana": {
+                            "type": "fruit",
+                            "properties": {
+                                "color": "yellow"
+                            }
+                        },
+                        "orange": {
+                            "type": "fruit",
+                            "properties": {
+                                "color": "orange"
+                            }
+                        }
+                    },
+                    "secure": []
+                }
+            },
+            "defaults": {
+                "fruit": "fruit.apple",
+                "banana": "fruit.banana"
+            },
+            "plugins": [
+                "@zowe/fruit-for-imperative"
+            ]
+        });
+
+        // Allocate the command processor
+        const processor: CommandProcessor = new CommandProcessor({
+            envVariablePrefix: ENV_VAR_PREFIX,
+            fullDefinition: SAMPLE_CMD_WITH_OPTS_AND_PROF, // `group action`
+            definition: {
+                name: "banana",
+                description: "The banana command",
+                type: "command",
+                handler: __dirname + "/__model__/TestArgHandler",
+                options: [
+                    {
+                        name: "boolean-opt",
+                        type: "boolean",
+                        description: "A boolean option.",
+                    },
+                    {
+                        name: "color",
+                        type: "string",
+                        description: "The banana color.",
+                        required: true
+                    }
+                ],
+                profile: {
+                    optional: ["banana"]
+                }
+            },
+            helpGenerator: FAKE_HELP_GENERATOR,
+            profileManagerFactory: FAKE_PROFILE_MANAGER_FACTORY,
+            rootCommandName: SAMPLE_ROOT_COMMAND,
+            commandLine: "",
+            promptPhrase: "dummydummy",
+            config:  ImperativeConfig.instance.config
+        });
+
+        // Mock the profile loader
+        (CommandProfileLoader.loader as any) = jest.fn((args) => {
+            return {
+                loadProfiles: (profArgs: any) => {
+                    return;
+                }
+            };
+        });
+
+        const parms: any = {
+            arguments: {
+                _: ["check", "for", "banana"],
+                $0: "",
+                [parm1Key]: parm1Value,
+                valid: true,
+                showInputsOnly: true,
+            },
+            silent: true
+        };
+        const commandResponse: ICommandResponse = await processor.invoke(parms);
+        expect(commandResponse.data.locations.length).toBeGreaterThan(0);
+        expect(commandResponse.data.optionalProfiles[0]).toBe(`banana`);
+        expect(commandResponse.data.requiredProfiles).toBeUndefined();
+        expect(commandResponse.data.profileVersion).toBe('v2');
+    });
+
+
+    it("should mask input value for a secure parm when --show-inputs-only flag is set", async () => {
+
+        // values to test
+        const parm1Key = `user`;
+        const parm1Value = `secret`;
+        const secure = `(secure value)`;
+
+        // Allocate the command processor
+        const processor: CommandProcessor = new CommandProcessor({
+            envVariablePrefix: ENV_VAR_PREFIX,
+            fullDefinition: SAMPLE_COMPLEX_COMMAND, // `group action`
+            definition: { // `object`
+                name: "banana",
+                description: "The banana command",
+                type: "command",
+                handler: __dirname + "/__model__/TestCmdHandler",
+                options: [
+                    {
+                        name: parm1Key,
+                        type: "string",
+                        description: "The first parameter",
+                    }
+                ],
+            },
+            helpGenerator: FAKE_HELP_GENERATOR,
+            profileManagerFactory: FAKE_PROFILE_MANAGER_FACTORY,
+            rootCommandName: SAMPLE_ROOT_COMMAND,
+            commandLine: "",
+            promptPhrase: "dummydummy"
+        });
+
+        const parms: any = {
+            arguments: {
+                _: ["check", "for", "banana"],
+                $0: "",
+                [parm1Key]: parm1Value,
+                valid: true,
+                showInputsOnly: true,
+            },
+            silent: true
+        };
+        const commandResponse: ICommandResponse = await processor.invoke(parms);
+        expect(commandResponse.data.commandValues[parm1Key]).toBe(secure);
+        expect(commandResponse.stderr.toString()).toContain(`Some inputs are not displayed`);
+    });
+
+    it("should not mask input value for a secure parm when --show-inputs-only flag is set with env setting", async () => {
+
+        // values to test
+        const parm1Key = `user`;
+        const parm1Value = `secret`;
+
+        process.env["test-cli_SHOW_SECURE_ARGS"] = "true";
+
+        // Allocate the command processor
+        const processor: CommandProcessor = new CommandProcessor({
+            envVariablePrefix: ENV_VAR_PREFIX,
+            fullDefinition: SAMPLE_COMPLEX_COMMAND, // `group action`
+            definition: { // `object`
+                name: "banana",
+                description: "The banana command",
+                type: "command",
+                handler: __dirname + "/__model__/TestCmdHandler",
+                options: [
+                    {
+                        name: parm1Key,
+                        type: "string",
+                        description: "The first parameter",
+                    }
+                ],
+            },
+            helpGenerator: FAKE_HELP_GENERATOR,
+            profileManagerFactory: FAKE_PROFILE_MANAGER_FACTORY,
+            rootCommandName: SAMPLE_ROOT_COMMAND,
+            commandLine: "",
+            promptPhrase: "dummydummy"
+        });
+
+        const parms: any = {
+            arguments: {
+                _: ["check", "for", "banana"],
+                $0: "",
+                [parm1Key]: parm1Value,
+                valid: true,
+                showInputsOnly: true,
+            },
+            silent: true
+        };
+        const commandResponse: ICommandResponse = await processor.invoke(parms);
+        expect(commandResponse.data.commandValues[parm1Key]).toBe(parm1Value);
+        expect(commandResponse.stderr.toString()).not.toContain(`Some inputs are not displayed`);
+
+        delete process.env["test-cli_SHOW_SECURE_ARGS"];
+    });
+
     it("should invoke the handler and process daemon response and then return success=true if the handler was successful", async () => {
         // Allocate the command processor
         const processor: CommandProcessor = new CommandProcessor({
@@ -1690,7 +1995,7 @@ describe("Command Processor", () => {
             commandLine: "",
             promptPhrase: "dummydummy"
         });
-        const commandResponse: ICommandResponse = await processor.help(new CommandResponse({silent: true}));
+        const commandResponse: ICommandResponse = await processor.help(new CommandResponse({ silent: true }));
         expect(commandResponse).toBeDefined();
         expect(commandResponse).toMatchSnapshot();
     });